requirements.txt
alpaca-trade-api==3.0.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.14.0
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
websockets>=9.0,<11.0
yfinance==0.2.36
ta-lib==0.4.28
pandas-ta==0.3.14b0
textblob==0.17.1
nltk==3.8.1
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0 